made additional method and formatting changes
