+++ hackathon_site/dashboard/frontend/src/api/helpers.ts
@@ -6,6 +6,18 @@ import {
     ReturnOrderInTable,
 } from "api/types";
 
+export const formatDateTime = (dateTimeString: string): string => {
+    const dateTime = new Date(dateTimeString);
+    const options: Intl.DateTimeFormatOptions = {
+        month: "long",
+        day: "numeric",
+        year: "numeric",
+        hour: "numeric",
+        minute: "numeric",
+        hour12: true,
+    };
+    return dateTime.toLocaleString("en-US", options);
+};
 export const teamOrderListSerialization = (
     orders: Order[]
 ): {
