+++ hackathon_site/dashboard/frontend/src/components/general/OrderTables/OrderTables.module.scss
@@ -70,6 +70,16 @@
     }
 }
 
+.chipPaddingRight {
+    position: relative;
+    left: 630px;
+}
+
+.chipPaddingLeft {
+    position: relative;
+    left: 629px;
+}
+
 .noWrap {
     white-space: nowrap;
 }
